requirements.txt
alembic==1.13.2
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
binaryornot==0.4.4
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.4
cycler==0.12.1
dacite==1.8.1
dj-database-url==2.3.0
Django==5.1.4
django-coreapi==2.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-Validates==0.3.0
Flask-WTF==1.2.1
fonttools==4.53.1
greenlet==3.0.3
gunicorn==22.0.0
htmlmin==0.1.12
idna==3.7
ImageHash==4.3.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
matplotlib==3.9.0
mdurl==0.1.2
multimethod==1.12
mysqlclient==2.2.4
networkx==3.3
numba==0.60.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.3
statsmodels==0.14.2
text-unidecode==1.3
tqdm==4.66.5
typeguard==4.3.0
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.2
visions==0.7.6
Werkzeug==3.0.3
wordcloud==1.9.3
WTForms==3.1.2
ydata-profiling==4.9.0
